# pkg/api/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "api.CreateSSHKeyRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "public_key": {
                        "type": "string"
                    }
                },
                "required": [
                    "name",
                    "public_key"
                ],
                "type": "object"
            },
...
            "api.ErrorResponse": {
                "properties": {
                    "error": {
                        "type": "string"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "api.StatusResponse": {
                "properties": {
                    "status": {
                        "type": "string"
                    }
                },
                "required": [
                    "status"
                ],
                "type": "object"
            },
...
            "types.SSHKey": {
                "properties": {
                    "created_at": {
                        "type": "integer"
                    },
                    "id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "public_key": {
                        "type": "string"
                    },
                    "user_id": {
                        "type": "string"
                    }
                },
                "required": [
                    "created_at",
                    "id",
                    "name",
                    "public_key",
                    "user_id"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "description": "Bearer JWT obtained from POST /api/login.",
                "in": "header",
                "name": "Authorization",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "VM lifecycle, networking, media and jobs API. Mutations returning 202 are asynchronous; inspect the returned job for completion.",
        "title": "Maco API",
        "version": "1.0"
    },
...
    "paths": {
...
        "/api/me/ssh-keys": {
            "get": {
                "description": "List the SSH public keys stored on the current account. These keys populate [[ .SSHKeys ]] in guest configurations when a VM is created.",
                "operationId": "listSSHKeys",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "items": {
                                        "$ref": "#/components/schemas/types.SSHKey"
                                    },
                                    "type": "array"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "listSSHKeys",
                "tags": [
                    "ssh-keys"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": true,
                    "transport": "json"
                }
            },
            "post": {
                "description": "Add an SSH public key to the current account. Supply a name and a complete public key such as ssh-ed25519 AAAA\u2026. Never submit a private key.",
                "operationId": "addSSHKey",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/api.CreateSSHKeyRequest",
                                "summary": "body",
                                "description": "Request"
                            }
                        }
                    },
                    "description": "Request",
                    "required": true
                },
                "responses": {
                    "201": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/types.SSHKey"
                                }
                            }
                        },
                        "description": "Created"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Bad Request"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "addSSHKey",
                "tags": [
                    "ssh-keys"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            }
        },
        "/api/me/ssh-keys/{id}": {
            "delete": {
                "description": "Delete an SSH public key from the current account by ID from listSSHKeys.",
                "operationId": "deleteSSHKey",
                "parameters": [
                    {
                        "description": "SSH key id",
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.StatusResponse"
                                }
                            }
                        },
                        "description": "OK"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unauthorized"
                    },
                    "404": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Not Found"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/api.ErrorResponse"
                                }
                            }
                        },
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "deleteSSHKey",
                "tags": [
                    "ssh-keys"
                ],
                "x-maco": {
                    "expose": true,
                    "readOnly": false,
                    "transport": "json"
                }
            }
        },
...
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "url": "/"
        }
    ],
...
}